requirements.txt
asgiref==3.2.7
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-extensions==2.2.9
django-shell-plus==1.1.7
djangorestframework==3.11.0
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
psycopg2-binary==2.8.5
django-auth-ldap==2.2.0
